Using proper code to measure time

+++ src/framemonitor.h
@@ -1,11 +1,13 @@
 #pragma once
 #include <vector>
+#include "common.h"
+
 class FrameMonitor
 {
     bool bShow = true;
     std::vector<float> values;
     int frame = 0;
-    double lastTime = 0.0;
+    u64 lastTime = 0;
 
 public:
     FrameMonitor();
